Fix cmdline runner test
